Use kws thresholds file
Reworked the pocketsphinx kws to use a thresholds file. This file
is in the format:

KEYWORD /1e+threshold/

where threshold is generally between -50 and 50. The higher the
threshold, the less likely Naomi is to think it hears the
keyword in random speech around it, and the more likely it is to
not hear the keyword when you do speak it.

I generally find the range between -10 and 10 works best.