Change default threshold
Based on my testing, the default threshold should be set much
higher than the previous default of 25.

My Naomi has been sitting on my desk for the last couple of months
in a large shared workspace. I have had the out audio going into
headphones to avoid annoying others, so I really haven't been using
it much, and any wakewords it does hear are mostly things it is
mishearing from the environment.

Here is a log of the number of times it has been hearing a wakeword
lately:

2021-06-01|41|MAGICVOICE
2021-06-01|203|NAOMI
2021-06-02|72|COMPUTER
2021-06-02|22|MAGICVOICE
2021-06-02|271|NAOMI
2021-06-03|120|COMPUTER
2021-06-03|16|MAGICVOICE
2021-06-03|327|NAOMI
2021-06-04|102|COMPUTER
2021-06-04|10|MAGICVOICE
2021-06-04|267|NAOMI
2021-06-05|1|COMPUTER
2021-06-05|2|NAOMI
2021-06-06|1|NAOMI
2021-06-07|51|COMPUTER
2021-06-07|7|MAGICVOICE
2021-06-07|141|NAOMI
2021-06-08|77|COMPUTER
2021-06-08|11|MAGICVOICE
2021-06-08|226|NAOMI

(note that the 5th was a Saturday and the 6th was a Sunday and
really nobody was in the office on those days).

It is interesting to note that it is much more likely to hear
Naomi than either of the other wake words, despite the fact that
MagicVoice (my default) is set to a threshold of 75, Computer is
set to a threshold of 200 and Naomi is currently set to 300.