Test failed
Tested to see if things would work with the image directory with no image file but it does not. Removing the logo toml. After this, you can add the commit change to your PR in the naomi-plugins repo, which I will then approve, and then the site will finally be able to build and be up to date once again.