Now compatible with new_pocketsphinx
I updated this to be compatible with my new_pocketsphinx pull
request for Naomi. Here are the major differences:

1) Uses the Phonetisaurus python plugin rather than calling the
phonetisaurus executable directly through subprocess. This means
that we no longer need to know which phonetisaurus executable we
are using.

2) No longer generating a language model. Turns out that when
using pocketsphinx you have a choice of using a language model,
a KeywordSearch file or a JSGF grammar file. Since we are using
a KeywordSearch file, we don't need the language model, so I'm
not generating it.

3) All files are downloaded/generated by the plugin. This plugin
no longer depends on anything installed with the pocketsphinx
plugin, so now these can be installed independently.

4) Moved and renamed the phonetisaurus fst file. The fst file
generated by phonetisaurus from the dictionary is now put
directly into the hmm directory and named g2p_model.fst
instead of being in the train subdirectory and named model.fst
since I'm going to need to retrain the model anyway due to
this model using lower case words.