Fixed compile_vocabulary
I had wanted to rename compile_languagemodel because we don't
actually need a languagemodel. The output of the function was a
list of all the words to put in the dictionary stored in a
variable called vocabulary. So I renamed it compile_vocabulary.

Unfortunately, there was already a "compile_vocabulary" function
which is used to compile the dictionary. This was causing an error,
so I have re-named the compile_languagemodel function to
compile_lexicon. It's still confusing, but it works now.